# openapi.yaml
openapi: 3.0.0
info:
  title: Kazi Connect API
  version: 1.2.0
  description: >
    Kazi Connect job listings and job applications for SettleIn students, plus accounts and
    employer tools for the FundiLink website. Endpoint numbers match ENDPOINT_LIST.md, and each
    endpoint names the need statement it serves. Endpoints marked with a lock need a login token:
    log in with Endpoint 11, then send the token as "Authorization: Bearer <token>".
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /api/jobs:
    get:
      summary: Returns active job/service listings, optionally filtered by county and category
      description: >
        Endpoint 1 (Statement 1): send only status=active to get every active listing.
        Endpoint 2 (Statement 2): also send county and category to get the active listings near a
        student's accommodation. The filters are query parameters on the jobs resource, not a
        separate path. An empty list means nothing matched.
      parameters:
        - name: status
          in: query
          required: true
          description: active returns jobs that can still be requested; closed returns jobs that are filled or withdrawn.
          schema:
            type: string
            example: active
        - name: county
          in: query
          required: false
          description: Endpoint 2 filter. Matches the job's county, ignoring upper and lower case.
          schema:
            type: string
            example: Nairobi
        - name: category
          in: query
          required: false
          description: Endpoint 2 filter. Matches the job's category, ignoring upper and lower case.
          schema:
            type: string
            example: accommodation
      responses:
        '200':
          description: The matching job listings. Empty list if none match.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
        '400':
          description: status is missing, or county or category was sent empty.

    # ENDPOINT 14 (same path as Endpoints 1 and 2)
    post:
      summary: Posts a new job (employers only)
      description: FundiLink Statement F2. The kind of work is saved in lower case. A new job starts active.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewJob'
      responses:
        '201':
          description: Job posted. The Location header points to the new job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: The body failed validation. Nothing is posted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: 'payRange.min must not be more than payRange.max'
        '401':
          $ref: '#/components/responses/NotLoggedIn'
        '403':
          $ref: '#/components/responses/Forbidden'

  #---ENDPOINT 3---
...
  /api/applications:
    post:
      summary: Creates a new application/service request from a SettleIn student
      description: >
        Statement 6. The job must exist and be active. Each call creates a new application, so this
        endpoint is not idempotent: sending the same body twice creates two applications.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewApplication'
      responses:
        '201':
          description: Application successfully created. The Location header points to the new application.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          description: >
            The body failed validation: a required field is missing, a field has the wrong type,
            a value is unusable (empty text, not a real date-time, a date in the past), a field
            is not allowed, or jobId does not match any job. Nothing is written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: 'Missing required field(s): preferredDate'
        '409':
          description: The job exists but is closed, so it is not accepting applications. Nothing is written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Job job_12350 is closed and is not accepting applications

  #---ENDPOINT 8 and ENDPOINT 9 (same path, different methods)---
...
  /api/me:
    get:
      summary: Returns the logged in user
      description: FundiLink Statement F1.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The logged in user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/NotLoggedIn'

  #---ENDPOINT 15---
  /api/me/jobs:
    get:
      summary: Lists the jobs the logged in employer posted
      description: FundiLink Statement F2. Newest first, each with how many applications are waiting.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The employer's jobs. Empty list if none posted yet.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployerJob'
        '401':
          $ref: '#/components/responses/NotLoggedIn'
        '403':
          $ref: '#/components/responses/Forbidden'

  #---ENDPOINT 19---
  /api/me/applications:
    get:
      summary: Lists the logged in student's applications
      description: >
        FundiLink Statement F5. Applications whose studentId is the student's account ID, newest
        first. Empty list if the student has not applied yet.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The student's applications.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Application'
        '401':
          $ref: '#/components/responses/NotLoggedIn'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'The token from Endpoint 10 or 11, sent as "Authorization: Bearer <token>".'
  responses:
    NotLoggedIn:
      description: No token was sent, or it is not valid, or it has expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: 'Your session has expired or is not valid. Log in again.'
    Forbidden:
      description: Logged in, but this account is not allowed to do this.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: 'You can only manage jobs you posted.'
  schemas:
    Job:
      type: object
      required: [id, title, category, county, area, duration, urgent, status, payRange]
      properties:
        id:
          type: string
          example: job_12345
        title:
          type: string
          example: Campus Housing Assistant
        category:
          type: string
          example: accommodation
        county:
          type: string
          example: Nairobi
        area:
          type: string
          description: The estate or neighbourhood within the county.
          example: Parklands
        duration:
          type: string
          description: How long the work takes, in plain words.
          example: Part time, 3 months
        urgent:
          type: boolean
          description: true when the employer needs someone quickly.
          example: true
        status:
          type: string
          enum: [active, closed]
          description: active means the job can still be requested; closed means it is filled or withdrawn.
          example: active
        payRange:
          type: object
          required: [min, max]
          properties:
            min:
              type: number
              example: 15000
            max:
              type: number
              example: 22000
...
    Application:
      type: object
      required: [applicationId, jobId, studentId, preferredDate, note, status]
      properties:
        applicationId:
          type: string
          example: app_1005
        jobId:
          type: string
          example: job_12345
        studentId:
          type: string
          example: stu_0042
        preferredDate:
          type: string
          format: date-time
          example: '2027-01-15T09:00:00Z'
        note:
          type: string
          description: Empty string when the student left no note.
          example: Available after 2pm on weekdays
        status:
          type: string
          enum: [pending, accepted, filled]
          example: pending
    NewApplication:
      type: object
      additionalProperties: false
      required: [jobId, studentId, preferredDate]
      properties:
        jobId:
          type: string
          minLength: 1
          description: Must match an existing, active job.
          example: job_12345
        studentId:
          type: string
          minLength: 1
          example: stu_0042
        preferredDate:
          type: string
          format: date-time
          description: When the student wants the service. Must be a real date-time in the future.
          example: '2027-01-15T09:00:00Z'
        note:
          type: string
          maxLength: 500
          example: Available after 2pm on weekdays
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: 'Missing required field(s): preferredDate'
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: string
          example: usr_1001
        name:
          type: string
          example: Demo Employer
        email:
          type: string
          format: email
          example: employer@kaziconnect.demo
        role:
          type: string
          enum: [student, employer]
          example: employer
...
    NewJob:
      type: object
      additionalProperties: false
      required: [title, category, county, area, duration, payRange]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 100
          example: Fix leaking kitchen sink
        category:
          type: string
          minLength: 2
          maxLength: 40
          description: The kind of work. Saved in lower case.
          example: plumbing
        county:
          type: string
          minLength: 2
          maxLength: 40
          example: Nairobi
        area:
          type: string
          minLength: 2
          maxLength: 60
          example: Westlands
        duration:
          type: string
          minLength: 2
          maxLength: 40
          example: About 2 hours
        urgent:
          type: boolean
          description: Leave out for false.
          example: true
        payRange:
          type: object
          additionalProperties: false
          required: [min, max]
          description: Pay in Kenyan shillings. min must not be more than max.
          properties:
            min:
              type: number
              minimum: 0
              maximum: 10000000
              example: 1500
            max:
              type: number
              minimum: 0
              maximum: 10000000
              example: 1500
    EmployerJob:
      type: object
      description: A Job, plus how many applications are waiting for a decision.
      required: [id, title, category, county, area, duration, urgent, status, payRange, pendingApplications]
      properties:
        id:
          type: string
          example: job_12345
        title:
          type: string
          example: Campus Housing Assistant
        category:
          type: string
          example: accommodation
        county:
          type: string
          example: Nairobi
        area:
          type: string
          example: Parklands
        duration:
          type: string
          example: Part time, 3 months
        urgent:
          type: boolean
          example: true
        status:
          type: string
          enum: [active, closed]
          example: active
        payRange:
          type: object
          required: [min, max]
          properties:
            min:
              type: number
              example: 15000
            max:
              type: number
              example: 22000
        pendingApplications:
          type: integer
          minimum: 0
          example: 1
...